src/main/java/com/julenka/molodec/userapplication/services/DeleteUserService.java
package com.julenka.molodec.userapplication.services;

import com.julenka.molodec.userapplication.domain.Role;
import com.julenka.molodec.userapplication.domain.User;
import com.julenka.molodec.userapplication.repository.RoleRepository;
import com.julenka.molodec.userapplication.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class DeleteUserService {

    @Autowired
    private RoleRepository roleRepository;
    @Autowired
    private UserRepository userRepository;


    public void deleteUser(User user) {
        roleRepository.deleteAll(roleRepository.findAllByUser(user));
        userRepository.delete(user);
    }
}

