src/test/java/com/julenka/molodec/userapplication/services/DeleteUserServiceTest.java
package com.julenka.molodec.userapplication.services;

import com.julenka.molodec.userapplication.domain.Role;
import com.julenka.molodec.userapplication.domain.User;
import com.julenka.molodec.userapplication.repository.RoleRepository;
import com.julenka.molodec.userapplication.repository.UserRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.ArgumentCaptor;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class DeleteUserServiceTest {

    @Mock
    private RoleRepository roleRepository;

    @Mock
    private UserRepository userRepository;

    @InjectMocks
    private DeleteUserService deleteUserService;

    private static final List<Role> ROLES = List.of(new Role());

    @BeforeEach
    void configureRoleRepository(){
        when(roleRepository.findAllByUser(any(User.class))).thenReturn(ROLES);
    }

    @Test
    void deleteUserAndHisRoles(){
        User user = new User();
        deleteUserService.deleteUser(user);
        verify(roleRepository).deleteAll(ROLES);
        ArgumentCaptor<User> userArgumentCaptor = ArgumentCaptor.forClass(User.class);
        verify(userRepository).delete(userArgumentCaptor.capture());
        User captured = userArgumentCaptor.getValue();
        assertEquals(user, captured);
    }

}